filled out documentation files and renamed covariance functions for consistency
