+++ docs/ForcingCovariance.rst
@@ -1,0 +1,11 @@
+.. _ForcingCovariance:
+
+**********************************
+The ``ForcingCovariance`` Class
+**********************************
+
+
+.. autoclass:: stat_fem.ForcingCovariance.ForcingCovariance
+    :members:
+
+    .. automethod:: __init__
