+++ docs/InterpolationMatrix.rst
@@ -1,0 +1,11 @@
+.. _InterpolationMatrix:
+
+**********************************
+The ``InterpolationMatrix`` Class
+**********************************
+
+
+.. autoclass:: stat_fem.InterpolationMatrix.InterpolationMatrix
+    :members:
+
+    .. automethod:: __init__
