+++ docs/LinearSolver.rst
@@ -1,0 +1,11 @@
+.. _LinearSolver:
+
+**********************************
+The ``LinearSolver`` Class
+**********************************
+
+
+.. autoclass:: stat_fem.LinearSolver.LinearSolver
+    :members:
+
+    .. automethod:: __init__
