+++ docs/ObsData.rst
@@ -1,0 +1,11 @@
+.. _ObsData:
+
+**********************************
+The ``ObsData`` Class
+**********************************
+
+
+.. autoclass:: stat_fem.ObsData.ObsData
+    :members:
+
+    .. automethod:: __init__
