+++ docs/assemble.rst
@@ -1,0 +1,9 @@
+.. _assemble:
+
+**********************************
+The ``assemble`` Module
+**********************************
+
+
+.. automodule:: stat_fem.assemble
+    :members:
