+++ docs/covariance_functions.rst
@@ -1,0 +1,9 @@
+.. _covariance_functions:
+
+***********************************
+The ``covariance_functions`` Module
+***********************************
+
+
+.. automodule:: stat_fem.covariance_functions
+    :members:
