+++ docs/estimation.rst
@@ -1,0 +1,9 @@
+.. _estimation:
+
+**********************************
+The ``estimation`` Module
+**********************************
+
+
+.. automodule:: stat_fem.estimation
+    :members:
