+++ docs/solving.rst
@@ -1,0 +1,9 @@
+.. _solving:
+
+**********************************
+The ``solving`` Module
+**********************************
+
+
+.. automodule:: stat_fem.solving
+    :members:
