+++ docs/solving_utils.rst
@@ -1,0 +1,9 @@
+.. _solving_utils:
+
+**********************************
+The ``solving_utils`` Module
+**********************************
+
+
+.. automodule:: stat_fem.solving_utils
+    :members:
